requirements.txt
customtkinter
tkinterdnd2
natsort
pillow
puremagic
pillow_heif
tqdm
pillow-avif-plugin
filetype
numpy
pydub
fleep
dotted_dict
pyinstaller
pyperclip